// server/src/models/schemas/Device.schema.ts
export interface DeviceType {
  _id?: ObjectId
  created_at?: Date
  updated_at?: Date
  uAParse?: {
    deviceUAParse_name: string
    deviceUAParse_cpu_architecture: string
    deviceUAParse_engine_name: string
    deviceUAParse_engine_version: string
    deviceUAParse_os_name: string
    deviceUAParse_os_version: string
    deviceUAParse_ua: string
    deviceUAParse_version: string
  }
  device?: {
    hostname: string
    platform: string
    cpuInfo: ObjectConstructor[][]
    fileInfo: number
    networkInfo: {
      Ethernet: ArrayConstructor[]
      'Loopback Pseudo-Interface 1': ArrayConstructor[]
      'Teredo Tunneling Pseudo-Interface': ArrayConstructor[]
    }
    osType: string
    osPlatform: string
    osArch: string
    userInfo: {
      uid: number
      gid: number
      username: string
      homedir: string
      shell: null
    }

    totalMemory: number
    freeMemory: number
    uptime: number
    osConstants: {
      UV_UDP_REUSEADDR: number
      dlopen: object
      errno: ObjectConstructor[]
      signals: ObjectConstructor[]
      priority: ObjectConstructor[]
    }
    is64Bit: boolean
    loadAvg: number[]
    isLittleEndian: boolean
  }
}